Split back into separate row, column and box modules
