+++ src/invent/ui/__init__.py
@@ -22,9 +22,10 @@ from ..i18n import _
 from .app import App
 from .core import Widget, Container, from_datastore
 from .utils import random_id, sanitize
-from .containers.gridbox import Column, Row
+from .containers.column import Column
 from .containers.grid import Grid
 from .containers.page import Page
+from .containers.row import Row
 from .widgets.audio import Audio
 from .widgets.button import Button
 from .widgets.checkbox import CheckBox
