+++ src/invent/ui/containers/box.py
@@ -1,0 +1,8 @@
+from ..core import Container
+
+
+class Box(Container):
+    def render(self):
+        element = super().render()
+        element.style.display = "grid"
+        return element
