+++ src/invent/ui/containers/column.py
@@ -1,0 +1,17 @@
+from .box import Box
+from .grid import set_position
+
+
+class Column(Box):
+    """
+    A vertical container box.
+    """
+
+    @classmethod
+    def icon(cls):
+        return '<svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 256 256"><path fill="currentColor" d="M104 32H64a16 16 0 0 0-16 16v160a16 16 0 0 0 16 16h40a16 16 0 0 0 16-16V48a16 16 0 0 0-16-16m0 176H64V48h40Zm88-176h-40a16 16 0 0 0-16 16v160a16 16 0 0 0 16 16h40a16 16 0 0 0 16-16V48a16 16 0 0 0-16-16m0 176h-40V48h40Z"/></svg>'  # noqa
+
+    def update_child(self, child, index):
+        child.element.style.setProperty("grid-column", 1)
+        child.element.style.setProperty("grid-row", index)
+        set_position(child.element, child.position)
